pyproject.toml
[tool.pytest.ini_options]
addopts = "--import-mode=importlib"

[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "3.11.*"
numpy = "2.0.0"
gpxpy = "1.6.2"
click = "8.1.7"
requests = "2.32.3"
pillow = "10.4.0"
pyexiftool = "0.5.6"
python-dotenv = "1.0.1"
gdal = "==3.8.5.*"

[tool.poetry.group.test.dependencies]
black = "24.4.2"
mypy = "1.10.1"
types-requests = "2.32.0.20240622"
isort = "5.13.2"
pylint = "3.2.5"
pytest = "8.2.2"
coverage = "7.5.4"
pytest-vcr = "1.0.2"
pytest-cov = "5.0.0"
robotframework = "7.0.1"
robotframework-seleniumlibrary = "6.5.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
